# API/medical-management-api/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Medical managment API",
    "description" : "Medical managment API for hospital staff and treatments managment",
    "contact" : {
      "email" : "matej.mikus96@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/NonActive/MedicalManagment/1.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  } ],
...
  "paths" : {
...
    "/treatment" : {
      "post" : {
        "tags" : [ "admins" ],
        "summary" : "Creates new treatment",
        "description" : "Stores new treatment",
        "operationId" : "createTreatment",
        "requestBody" : {
          "description" : "Treatment model to create",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Treatment"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "treatment created"
          },
          "400" : {
            "description" : "treatment object invalid"
          },
          "409" : {
            "description" : "the provided treatment already exists"
          }
        }
      }
    },
    "/treatment/{treatmentId}" : {
      "put" : {
        "tags" : [ "admins" ],
        "summary" : "Updates existing treatment",
        "operationId" : "updateTreatment",
        "parameters" : [ {
          "name" : "treatmentId",
          "in" : "path",
          "description" : "pass the id of the particular treatment",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "description" : "Treatment model to update",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Treatment"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "treatment updated"
          },
          "400" : {
            "description" : "treatment object invalid"
          }
        }
      },
      "delete" : {
        "tags" : [ "admins" ],
        "summary" : "Deletes specific treatment",
        "operationId" : "deleteTreatment Use this method to delete the specific treatment",
        "parameters" : [ {
          "name" : "treatmentId",
          "in" : "path",
          "description" : "pass the id of the particular treatment",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Item deleted"
          },
          "404" : {
            "description" : "Treatment with such ID does not exists"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "SimpleEmployee" : {
        "required" : [ "degree", "firstName", "id", "lastName" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "number",
            "example" : 1
          },
          "firstName" : {
            "type" : "string",
            "example" : "John"
          },
          "lastName" : {
            "type" : "string",
            "example" : "Antal"
          },
          "degree" : {
            "type" : "string",
            "example" : "MUDr."
          }
        },
        "example" : {
          "firstName" : "John",
          "lastName" : "Antal",
          "degree" : "MUDr.",
          "id" : 1
        }
      },
...
      "Treatment" : {
        "required" : [ "date", "description", "id", "mainDoctorId", "otherDoctorsId", "room", "title", "type" ],
        "properties" : {
          "id" : {
            "type" : "number",
            "example" : 596
          },
          "title" : {
            "type" : "string",
            "example" : "Left knee surgerys"
          },
          "type" : {
            "type" : "number",
            "example" : 5
          },
          "description" : {
            "type" : "string",
            "example" : "Lorem ipsum and so on"
          },
          "mainDoctorId" : {
            "type" : "number",
            "example" : 1
          },
          "mainDoctorObject" : {
            "$ref" : "#/components/schemas/SimpleEmployee"
          },
          "otherDoctrsId" : {
            "type" : "array",
            "example" : [ 2, 3 ],
            "items" : {
              "type" : "number"
            }
          },
          "otherDoctorsObject" : {
            "type" : "array",
            "example" : [ {
              "id" : 2,
              "firstName" : "Mark",
              "lastName" : "Zuckerberg",
              "degree" : "MUDr"
            }, {
              "id" : 3,
              "firstName" : "Mocin",
              "lastName" : "Gejbr",
              "degree" : "MUDr"
            } ],
            "items" : {
              "type" : "object"
            }
          },
          "roomNumber" : {
            "type" : "string",
            "example" : "527/8 surgery room"
          },
          "date" : {
            "type" : "string",
            "format" : "date-time",
            "example" : "2017-07-21T17:32:28Z"
          }
        },
        "example" : {
          "date" : "2017-07-21T17:32:28Z",
          "otherDoctorsObject" : [ {
            "id" : 2,
            "firstName" : "Mark",
            "lastName" : "Zuckerberg",
            "degree" : "MUDr"
          }, {
            "id" : 3,
            "firstName" : "Mocin",
            "lastName" : "Gejbr",
            "degree" : "MUDr"
          } ],
          "mainDoctorObject" : {
            "firstName" : "John",
            "lastName" : "Antal",
            "degree" : "MUDr.",
            "id" : 1
          },
          "roomNumber" : "527/8 surgery room",
          "description" : "Lorem ipsum and so on",
          "id" : 596,
          "title" : "Left knee surgerys",
          "type" : 5,
          "otherDoctrsId" : [ 2, 3 ],
          "mainDoctorId" : 1
        }
      },
...
    }
  }
}